code: separated file with class
